phoree-chatbot/requirements.txt
groq
fastapi
uvicorn
python-dotenv 